+++ qrAttendanceSheet/__pycache__/apps.cpython-39.pyc
