+++ qrAttendanceSheet/__pycache__/models.cpython-39.pyc
