+++ qrAttendanceSheet/urls.py
@@ -15,17 +15,17 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from django.conf import settings
-from django.conf.urls.static import static
-from django.urls.conf import include
+# from django.conf import settings
+# from django.conf.urls.static import static
+from .views import home_view, register, student
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('', include('qrAttendanceSheet.urls')),
-
+    path('', home_view, name='home'),
+    path('register/<int:obj>', register, name='register'),
+    path('students/', student, name='students'),
 ]
 
-urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+# urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
