+++ qrc_proj/settings.py
@@ -37,8 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-
-    'websites',
+    'qrAttendanceSheet',
 ]
 
 MIDDLEWARE = [
