--- websites/forms.py
@@ -1,10 +1,0 @@
-from django.forms import ModelForm
-from .models import Student
-
-
-class StudentForm(ModelForm):
-
-    class Meta:
-        model = Student
-
-        fields=['website', 'student']
